fefs

+++ worksheet/portfolio/salary.c
@@ -25,7 +25,7 @@
    }
    float take_home= sal_ni-tax_con;
     // Use only these print statement with appropriate formatting and variable names
-   printf("Salary £%d\n",(int)salary);
+   printf("Salary £%.2f\n",salary);
    printf("NI contribution £%.2f\n", ni_con);
    printf("Tax contribution £%.2f\n", tax_con);
    printf("Take home salary £%.2f\n", take_home);
